Separate Python "dis" functions...
into its own file/module.
